+++ ChangeLog.markdown
@@ -1,6 +1,12 @@
 Changelog for Imbo
 ==================
 
+Imbo-2.x.x
+----------
+__N/A__
+
+* #434: Custom models for group(s) and access rule(s) (Christer Edvartsen)
+
 Imbo-2.0.0
 ----------
 __2016-02-28__